Implement custom training loop
